requirements.txt
torch==1.8.0
Pillow~=8.3.1
numpy~=1.21.2
torchvision~=0.11.2
matplotlib~=3.4.3